% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptivos.R
\name{descriptivos}
\alias{descriptivos}
\title{Descriptivos}
\usage{
descriptivos(x, na.rm = TRUE, trim = 0.1)
}
\arguments{
\item{x}{Conjunto de datos}

\item{na.rm}{Omitir valores perdidos}

\item{trim}{Valor de la media recortada}
}
\value{
Una tabla con los valores
}
\description{
Devuelve los estadísticos descriptivos habituales.
}
\examples{
x <- data.frame(a=sample(1:10, 100, replace =TRUE),
                  b=sample(1:10, 100, replace =TRUE),
                  c=sample(1:10, 100, replace =TRUE))
descriptivos(x)
}
