% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forma.R
\name{est.forma}
\alias{est.forma}
\title{Skewness and kurtosis}
\usage{
est.forma(datos, na.rm = TRUE, sktype = "fisher", mardia = TRUE, plot = FALSE)
}
\arguments{
\item{datos}{Data frame with data, variables in columns}

\item{na.rm}{Remove the NA}

\item{sktype}{The skewness formula: 'fisher' and 'pearson'}

\item{mardia}{Include the Mardia's skewness and kurtosis (multivariate)?}

\item{plot}{Plot the multivariate Q-Q plot}
}
\value{
A list with data frames: Forma are the values of variables and Multivariado the values of Mardia's skewness and kurtosis
}
\description{
Returns the simple and multivariate skewness and kurtosis. Results in Spanish
}
\examples{
x <- data.frame(a=sample(1:10, 100, replace =TRUE),
                  b=sample(1:10, 100, replace =TRUE),
                  c=sample(1:10, 100, replace =TRUE))
est.forma(x)

}
