% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moda.R
\name{moda}
\alias{moda}
\title{Mode of a vector or data frame}
\usage{
moda(x)
}
\arguments{
\item{x}{The vetor or data frame}
}
\value{
The mode as a vector or in a data frame. If there are multiples modes, it return a list with the variables with multiples modes.
}
\description{
If the input is a data frame, it calcule the mode of each column. If there are many modes, only returns the first.
}
\examples{
x <- data.frame(A = c(1,1,2,3), B=c(2,2,3,4))
moda(x)
}
